Flexbox container now takes full width of screen
the full width for the commputer screen is 100vh so I had to specify that exact value in each flexbox container section. I have a feeling I could make this cleaner to look at. Not only that but if I decide to come back than I could work on how the page would look on a smaller screen.